src/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
alembic=1.4.1=pypi_0
argh=0.26.2=py39haa95532_0
attrdict=2.0.1=pypi_0
azure-core=1.10.0=pypi_0
azure-storage-blob=12.7.1=pypi_0
ca-certificates=2021.1.19=haa95532_0
cached-property=1.5.2=pypi_0
certifi=2020.12.5=pypi_0
cffi=1.14.4=pypi_0
chardet=4.0.0=pypi_0
click=7.1.2=pypi_0
cloudpickle=1.6.0=pypi_0
configparser=5.0.1=pypi_0
cryptography=3.3.1=pypi_0
cycler=0.10.0=pypi_0
databricks-cli=0.14.1=pypi_0
dataclasses=0.6=pypi_0
datetime=4.3=pypi_0
dirsync=2.2.5=pypi_0
docker=4.4.1=pypi_0
entrypoints=0.3=pypi_0
flask=1.1.2=pypi_0
gitdb=4.0.5=pypi_0
gitpython=3.1.12=pypi_0
idna=2.10=pypi_0
isodate=0.6.0=pypi_0
itsdangerous=1.1.0=pypi_0
jinja2=2.11.2=pypi_0
kiwisolver=1.3.1=pypi_0
mako=1.1.4=pypi_0
markupsafe=1.1.1=pypi_0
matplotlib=3.3.4=pypi_0
minio=6.0.2=pypi_0
mlflow=1.13.1=pypi_0
monai=0.4.0=pypi_0
msrest=0.6.19=pypi_0
mysql-connector-python=8.0.23=pypi_0
numpy=1.19.5=pypi_0
oauthlib=3.1.0=pypi_0
opencv-python=4.5.1.48=pypi_0
openssl=1.1.1i=h2bbff1b_0
pandas=1.2.1=pypi_0
pathtools=0.1.2=py_1
pillow=8.0.1=pypi_0
pip=20.3.3=py39haa95532_0
prometheus-client=0.9.0=pypi_0
prometheus-flask-exporter=0.18.1=pypi_0
protobuf=3.14.0=pypi_0
pycparser=2.20=pypi_0
pyparsing=2.4.7=pypi_0
python=3.9.1=h6244533_2
python-dateutil=2.8.1=pypi_0
python-editor=1.0.4=pypi_0
pytz=2020.5=pypi_0
pywavelets=1.1.1=pypi_0
pywin32=227=pypi_0
pyyaml=5.3.1=py39h2bbff1b_1
querystring-parser=1.2.4=pypi_0
requests=2.25.1=pypi_0
requests-oauthlib=1.3.0=pypi_0
setuptools=51.3.3=py39haa95532_4
six=1.15.0=pypi_0
smmap=3.0.4=pypi_0
sqlalchemy=1.3.22=pypi_0
sqlite=3.33.0=h2a8f88b_0
sqlparse=0.4.1=pypi_0
tabulate=0.8.7=pypi_0
torch=1.7.1+cpu=pypi_0
torchaudio=0.7.2=pypi_0
torchvision=0.8.2+cpu=pypi_0
typing-extensions=3.7.4.3=pypi_0
tzdata=2020f=h52ac0ba_0
urllib3=1.26.3=pypi_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
waitress=1.4.4=pypi_0
watchdog=1.0.2=pypi_0
websocket-client=0.57.0=pypi_0
werkzeug=1.0.1=pypi_0
wheel=0.36.2=pyhd3eb1b0_0
wincertstore=0.2=py39h2bbff1b_0
yaml=0.2.5=he774522_0
zlib=1.2.11=h62dcd97_4
zope-interface=5.2.0=pypi_0
